Update README for new start procedure
Updated README to note the new start scripts that are available without requiring manual directory manipulation.

In the future, it would be nice to orchestrate this with Lerna or something similar.

For now, this feels sufficient. Particularly because the DB is still so finicky.

Need to containerize eventually with Docker, etc.